+++ CHANGELOG.md
@@ -11,10 +11,10 @@
  - #239 Add debug option in the service configuration file
  - #243 Add info about session management (README)
  - #230 Check missing schemas from local repo
+ - #251 Modify Event Subscription URI
 
 # New Redfish resources
  - EventService
- - EventSubscription
  - EventDestinationCollection
  - EventDestination